Add Dockerfile
This commit adds a Dockerfile to the repository containing all of the
dependencies necessary to run/develop Gematria. We have found this
pretty useful for the ml-compiler-opt repository as it allows for
new users to quickly experiment without having to deal with
"works on my machine errors" assuming they already have Docker
somewhat integrated into their workflow.