Added code for importing training data from LLVM and BHive CSV files.
Added the LLVM framework as a dependency and functions and classes
needed to convert LLVM data structures (llvm::MCInst) to Gematria data
structures. This is currently used mainly to implement the BHive
importer, but it can be used for building importers also from other
formats (e.g. from within an LLVM pass, or other binary analysis tools).

A tool that imports basic blocks from compiled ELF binaries based on
this code is coming in the near future.

Bonus changes:
- added a bibtex of the GRANITE paper to README.md.

PiperOrigin-RevId: 532481549