Format Python Code with Yapf
Now that yapf 0.40 (or 0.44?) has been released, we can use it to format
the gematria repository (if that's the formatter we decide to end up
going with). This does require a significant amount of initial
reformatting (but no open source tool that I've tested with perfectly
matches the style of the repository in its current state.

Add a note to the CONTRIBUTING file describing how to use yapf for
formatting.