Update Dockerfile after dependency installation change
Recently we moved from declaring an entire transitive closure in
requirements.txt (and generating that with Bazel) to declaring just the
top-level packages along with version information in requirements.in.
This change brought with it a different set of commands for setup which
breaks the existing Dockerfile. This patch fixes the Dockerfile to match
the new setup.