Only run push CI on main branch
Currently the github actions workflow runs whenever something is pushed
to a branch or against any pull request. This means that if a branch is
pushed to the main google/gematria repository and a PR is opened against
google/gematria:main, there will be duplicate jobs, one for the push
event and one for the pull request event. This patch fixes this behavior
by restricting the push event to the main branch.