Fix python formatting CI job
For some reason, using an older version of pyink causes odd failures in
the CI complaining about missing functions in a vendored dependency.
Upgrading to the most recent version fixes the issue. Maybe should try
and make this more hermetic at some point with a lockfile or something,
but for now this is a simple fix.