Unmap address 0x800000000 in the child process.
This is the address which will turn up when reading from newly mapped sections,
due to how we initialize them. If a block dereferences this pointer, we want it
to segfault, so that we can detect the read. On my machine it's unmapped, so all
is well. On some machines, it isn't.

Really we'd like to unmap everything we can, as blocks can access arbitrary
addresses. But this is difficult to do without hitting things which turn out to
be necessary and break things in weird ways when unmapped. But we can at least
easily fix this one issue with a single targetted munmap.

PiperOrigin-RevId: 596433115