Add exegesis BB annotator (#30)
This patch adds in an alternative to the existing find_accessed_addrs
infrastructure that uses llvm-exegesis as a backend. We believe this
will more closely match the execution environment of the benchmarking
environment and should avoid some of the pitfalls of the existing memory
annotation infrastructure (at the current expense of speed).