Make exegesis annotator work with mappings that aren't page aligned
This patch makes the exegesis annotator work with mappings that are not
page aligned. Currently if a segfault occurs at an address that isn't
page aligned, Exegesis will try and map an address at that address and
then the mmap call will silently fail as there is no error handling
wired up for that and it will segfault again, thus creating a loop
within the annotator.