Change constants in BHive conversion script
This patch adjusts the constants in the BHive conversion script to the
values used within the original BHive paper. Anecdotally, these
constants have also seemed to work better than the original ones. In
addition, this patch prefixes the memory value with zeroes so that
llvm-exegesis is able to assume the correct bit width.