Make exegesis conversion script use appropriate register class (#37)
This patch fixes the register class that the BHive to Exegesis
conversion script uses. Currently it is only pulling in registers that
don't have a REX encoding, which doesn't even include R9-R15. This patch
fixes that behavior by using the more generic LLVM 64-bit GPR Register
class, but specifically looking at registers that don't require a REX2
encoding, as there is no hardware in the wild yet that supports APX.