Add JSON output option to BHive converter (#38)
This patch adds a JSON output option to the BHive converter. This makes
it significantly easier to implement other scripts down the line that
ingest this data. This also cuts down the number of inodes that a large
data set will use by a significant amount, which can be a problem on
some file systems.