Make --blocks-per-json-file unsigned
Before this patch, --blocks-per-json-file was setup as unsigned in the
flag definitions, but the value was assigned to a normal int later on.
This resulted in the check that the value is <= triggering with the
default value of the maximum unsigned 32 bit integer. This patch fixes
that by assigning the value of the flag to an unsigned value rather than
a normal int.