Add progress reporting flag to annotator script
This patch adds a progress reporting flag to the annotator script. This
can be quite useful when using a slower implementation of
FindAccessedAddrs like FindAccessedAddrsExegesis to get a gauge on what
sort of progress is being made and to show that the application isn't
stuck somewhere when processing a large number of blocks.