Only add definitions for used registers
This patch makes the BHive to Exegesis conversion script only include
definitions for registers needed by the snippet rather than a random
set. This makes the outputs cleaner and will help keep things clean
later on when setting a loop register (and also make information
available to determine if/where there are conflicts).