Enable Bazel tests in CI
This patch enables running the Bazel tests in CI which allows us to
automatically catch regressions and more easily ensure that PRs aren't
breaking anything/the additional (tested) functionality is working
properly. This patch doesn't enable anything using perf_counters to be
tested in CI as Github Actions VMs don't have access to performance
counters. This effect is achieved by creating a bazel test tag and
excluding it on the bazel test invocation.