Improve --report_progress_every flag
Currently, the --report_progress_every flag, even with the default
value, will always report progress on the first block. This is not the
intended behavior, as no output should be given with the default value.
This patch fixes that behavior and also adds a test to ensure that the
flag functions as intended in the default value case and in the
value-provided case.

Not really super critical to functionality, but a little bit annoying to
have this.