Remove outdated comment
This patch removes the comment specifying that the call to
GetAccessedAddrs only got the first segfault rather than iteratively
looking through each one. This isn't true in the Exegesis case. It is
true in the current version for the fast annotator, but I believe this
has already been fixed internally and just needs to be pushed out into
the open.