Make blocks_per_json_file flag not output extra files (#63)
Currently, the logic in convert_bhive_to_llvm_exegesis_input will create
extra (empty) files as the output at the end runs unconditionally rather
than when there are actually blocks to output. This means that if we
have a number of blocks that is an exact multiple of the value in
blocks_per_json_file, we end up getting no blocks after the loop, which
means we output an empty JSON file.

Fixes #61.