Make exegesis annotator error out when remapping an address
Currently, there is a failure condition seen in some blocks where the
annotator finds a segfault, tells exegesis to map a page there, and then
the block fails again at the same address. This needs some debugging,
but currently it will be helpful if we can just log a warning message
and skip this basic block.