Change default annotator implementation to exegesis
This patch changes the default annotator implementation to exegesis.
This is due to the fast annotator deadlocking when used with future
patches that utilize threding to speed up the annotation process on
large datasets. This casues tests to randomly deadlock with the fast
annotator. Instead of manually changing each individual test, just
adjust the default.