Refactor find_accessed_addrs to use BasicBlockUtils
This patch refactors FindAccessedAddrs to use the BasicBlockUtils
library for getting the registers used in a block. This patch also does
some other refactorings to make the format used in find_accessed_addrs
more consistent with what is used elsewhere in the code base,
particularly changing the used register set to a vector of
registers/values.