Fix sort comparator in extract_bbs_from_obj (#144)
The comparator function for sorting basic blocks in extract_bbs_from_obj
previously compared the start address of the LHS to the size of the RHS
rather than comparing the start addresses. This patch fixes this
behavior. This makes the code correct (or at least fixes one correctness
issue) and fixes a libc++ assertion (when assertions are enabled).