Add date to bazel cache
This patch adds the current date to the bazel cache so that a new cache
will be stored every day rather than whenever the hashed files change.
This should significantly improve the build times for bazel as cache
hits should increase.

More improvements could be made in the future like adding the branch
name and restoring on main if there is no cache hit for the specific
branch.