Fix twine-related UB in convert_bhive_to_llvm_exegesis_input
According to the twine documentation, twine assignment when using
concatenation can lead to undefined behavior. This was leading to test
failures for convert_bhive_to_llvm_exegesis_input when running tests
with -c opt. This patch fixes this behavior by converting the twine to a
string and storing it in a std::string rather than trying to store it as
a twine.