Add ir_utils testing module (#156)
This patch adds the ir_utils testing module which contains some
functionality that is necessary for testing python infrastructure that
deals with IR in a clean manner, particularly beam pipelines. The
ir_utils module simply uses subprocess to call out to the LLVM utilities
like llvm-as and llvm-dis, which are listed as data dependencies for the
library.