Fix types and rename variables for clarity
Originally, the optimization pass list in OptimizeModules was typed as a
Sequence[Sequence[str]], with the intention being that each individual
pass would be a separate string. This doesn't make sense with how opt
takes in pass pipelines, which was why the typing information was
incorrect by the time the code landed. This patch changes the typing
information to be correct and renames the variables to hopefully provide
more clarity.