Fix tuple capitalization in benchmark_bbs_lib
This patch fixes capitalization of tuple in benchmark_bbs_lib. In a
patch update, we switched from Tuple to tuple (import from typing to the
standard python builtin), but I forgot to update one of the types, which
causes build failures locally.

Interestingly enough, the build failures do not occur in CI, which
probably should be investigated.