Use LLVMState helper for getting register index (#209)
This patch replaces our direct search of the register name to number
mapping to a function now included in LLVMState that handles that for
us, leaving us to solely deal with the error handling. Necessary now
that the accessor for the register name to number mapping will be
deleted soonish.