Change unittest.TestCase assertLen to assertEqual
unittest.TestCase has no documented assertLen function. I am not
entirely sure how this passes the tests, but apparently it does. Upon
importing into google though, we get an error due to the missing
function. This patch changes the test to use assertEqual, which is
documented as being a function on unittest.TestCase.