Add extra debugging for SIGILL.
This just dumps the 8 instruction bytes at the point at which we got the
SIGILL. You can then disassemble them manually to find which instruction
it failed on. This isn't super convenient (8 bytes is a fairly arbitrary
number to read), but it's easy to do and makes debugging somewhat
easier.