Support vector registers in FindAccessedAddrs (#234)
Support vector registers in FindAccessedAddrs

We only support 8 byte initial values, which are broadcast across the register.
To avoid introducing unnecessary uarch dependencies, we find the largest vector
register size used in the snippet, and only copy into that. Similarly, we detect
whether the 16 extended registers introduced in AVX-512 are used, and only copy
into them if so.

If no vector registers are used at all, we will skip over all instructions
touching vector registers to avoid as much of the overhead as possible.

Also add some extra debugging for SIGILL.

This just dumps the 8 instruction bytes at the point at which we got the
SIGILL. You can then disassemble them manually to find which instruction
it failed on. This isn't super convenient (8 bytes is a fairly arbitrary
number to read), but it's easy to do and makes debugging somewhat
easier.