Add GetVocab DoFn in compile_modules_lib
This patch adds an Apache Beam DoFn that gets the vocab from hex strings
of basic blocks. This is intended to be used in the annotation pipeline
to emit a vocab file along with the annotated blocks rather than having
to manually construct it later without the benefit of all the
parallelism.