Enable running annotations importer test internally
This patch fixes annotating_importer_test so that it can be run
internally as the bazel runfiles rule only kind of exists. We instead
just get the value from the TEST_SRCDIR environment variable and then
append the object file/perf data file path to it to get the data that we
need.