Get LLVM Binary paths in setup instead of init
This patch moves the code getting the LLVM binary paths to the beam DoFn
setup function rather than __init__. This is necessary as some
implementations of get_llvm_binary_path require extracting the binary
from an archive to a temporary directory on the local machine. Putting
this in __init__ does not guarantee that this happens on anything other
than the machine running it (one) whereas running it in setup ensures
the appropriate setup takes place for each worker instance.