Fix use after move to get exegesis annotator working in pipeline (#249)
This patch fixes a user after move in BHiveToExegesis/related classes where we would create a LLVMState, pass it into a couple other constructors to create things that would store a reference to LLVMState, and then promptly move it to call the BHiveToExegesis constructor. This resulted in use after moves, which were ultimately caught by ubsan. This patch fixes that behavior by encapsulating LLVMState within a unique_ptr to prevent the problems.

The exegesis annotator is also enabled in the compile_modules invocation script and test coverage is added now that it works.