Add support for df register in basic_block_utils (#253)
Certain instructions implicitly use the df register, which means we need
to support it for the exegesis annotator so that we do not get MC
verification errors. Additionally, it is good to explicitly define the
state to ensure we have deterministic execution. This patch adds support
for the df register by explicitly returning eflags as the super register
as LLVM does not seem to explicitly model df as a subregister of eflags.