Add benchmark success blocks counter to benchmarking pipeline (#256)
This patch adds a counter to track the number of blocks that have
succeeded. This is quite a useful metric to look at on the status page
for the Beam executor to track how fast things are progressing and how
far along things are.

Also actually increment the failed blocks counter because apparently I
forgot to do that in the last patch.