Add option to skip blocks with no loop register during annotation
This patch adds an option to the compile_modules pipeline to skip
emitting blocks that do not have any loop register attached to them
because they use all the GPRs. This is necessary when we want to
benchmark in loop mode.

This is intended to fix an issue where when running the benchmarking
pipeline (which is currently setup to only run in loop mode), we just
pass MCRegister::NoRegister, which causes a machine code verification
error as that is not valid. I will open another patch to fix that at
some point.