Add default benchmark scheduler
This patch adds a default benchmark scheduler. This scheduler assumes
the worker performing the benchmarking has three hyperthreads available
inside its CPU mask, with two of them being part of a pair (from the
same core). It then selects one core from the pair to be used for the
benchmarking subprocess, and selects the core outside of the pair to be
used to run the main exegesis/python program. Verification is also added
to ensure that nothing external to the program changes the process mask,
or at least to catch it early if it does happen.

Reviewers: orodley, ondrasej

Reviewed By: ondrasej

Pull Request: https://github.com/google/gematria/pull/271