Add function to get proto from hex and throughput value
This patch adds a new function to bhive_importer to directly create a
BasicBlockWithThroughputProto from a basic block in hex format and its
associated throughput value. No unittests are added on the C++ side as
this is a refactoring. A unit test is added on the python side to test
the bindings for the new function. This is intended to be used in the
benchmarking pipeline where formatting everything as a CSV before
passing it into the bhive importer makes little sense.

Reviewers: ondrasej, orodley

Reviewed By: ondrasej

Pull Request: https://github.com/google/gematria/pull/274