Ignore PerfCounterNotFullyEnabled in Exegesis Annotator
A recent LLVM commit makes llvm-exegesis return a
PerfCounterNotFullyEnabled error if the perf counter was not scheduled
on the CPU for the entire duration of the snippet measurement. We can
ignore this within the exegesis annotator because it implies the snippet
has executed successfully (as it can only happen at the end) and we do
not care about performance measurements within the annotator.

Reviewers: virajbshah, ondrasej, orodley

Reviewed By: orodley

Pull Request: https://github.com/google/gematria/pull/300