Make Exegesis Benchmark Retry Benchmarking Upon Perf Counter Failure
This patch makes exegesis_benchmark_lib retry benchmarking if the perf
counter is not enabled the entire time. This is part of the fix for zero
performance counter values on internal machines. Currently this
reassembles the snippet everytime a failure occurs, which could incur a
performance penalty. Given the interference patterns we have seen so far
though, this should not be a major issue. If this part of the code ends
up becoming a bottleneck, we can readjust later. The ideal place to do
this performance-wise is inside the BenchmarkRunner implementation
itself, but I want to avoid adding complexity there if we can avoid it.

Reviewers: orodley, virajbshah, ondrasej

Reviewed By: ondrasej

Pull Request: https://github.com/google/gematria/pull/301