Fix flaky annotation test (#303)
* Fix flaky annotation test

One of the tests in find_accessed_addrs_exegesis_test checks that if we
try and annotate the same address multiple times we bail out. It does
this by relying on the fact that exegesis does not error out on mmap
failures and by trying to annotate a snippet accessing memory above the
user space virtual memory ceiling. This is ceiling is higher on newer
platforms with five level page tables, so adjust it to be higher than
the new ceiling.

https://docs.kernel.org/5.10/x86/x86_64/mm.html

* feedback