Disable tests relying on TF1 behavior
As part of the conversion to TF2, this ptch disables all tests that rely
on TF1 behavior by calling disable_v2_behavior(). This is so that the
functionality can be incrementally ported to TF2 in separate patches
without having to worry about dependencies. The tests will be
selectively reenabled as the functionality is ported.

Reviewers: ondrasej

Reviewed By: ondrasej

Pull Request: https://github.com/google/gematria/pull/278