Port main_function to TF2
This patch ports the main_function library to TF2. Quite a bit of code
had to be adjusted to use newer APIs that are not incompatible with TF2.
Some functionality was completely removed like exporting graph defs as
it makes significantly less sense with TF2 if it is even supported.

Reviewers: ondrasej

Reviewed By: ondrasej

Pull Request: https://github.com/google/gematria/pull/285