Misc fixes for google import
This patch contains some misc fixes most related to type checking that
enable us to import gematria into google after the TF2 migration. Most
are pretty standard. We disable an attribute error on
run_continuous_evaluation mostly to silence it for now as tf_slim is not
really compatible with TF2 and needs to be rewritten.