Refactor loss_utils to be idiomatically TF2
This patch refactors loss_utils to be more TF2 like. The main impact of
this is that LossComputation now inherits from
tf.experimental.ExtensionType which implies everything about the class
has to be setup in the constructor. This enables returning a
LossComputation object from a function decorated with tf.function, which
enables us to wrap the entire training step in tf.function.

Reviewers: ondrasej, orodley, virajbshah

Reviewed By: ondrasej

Pull Request: https://github.com/google/gematria/pull/343