Use if chains instead of match case to avoid TF warnings (#357)
TF will issue warnings around match case statements. We can annotate
this with a function decorator to prevent autograph conversion, but the
performance implications of this, especially for JIT compilation, are
unclear. Use constructs that are supported even if they are slightly
less clean/maintainable to get rid of the warning and avoid any
performance penalties.