Misc pytype fixes for import (#360)
* Misc pytype fixes for import

There are a couple new pytype issues introduced by recent commits. This
patch fixes them. Namely the percentiles is expected to be a list of
lists rather than just a list and _unwrap_if_tensor does not currently
report returning that. We can also have None edge/node/global output
sizes in GraphNetworkLayer, so make sure to support that as well.

* formatting